+++ zcash_client_memory/src/testing/pool/sapling.rs
@@ -158,3 +158,9 @@ fn scan_cached_blocks_detects_change_notes() {
 fn scan_cached_blocks_detects_spends_out_of_order() {
     testing::pool::scan_cached_blocks_detects_spends_out_of_order::<SaplingPoolTester>()
 }
+
+#[test]
+#[ignore] //FIXME
+fn receive_two_notes_with_same_value() {
+    testing::pool::receive_two_notes_with_same_value::<SaplingPoolTester>()
+}