+++ zcash_client_sqlite/src/wallet/sapling.rs
@@ -704,4 +704,9 @@ pub(crate) mod tests {
     fn zip315_can_spend_inputs_by_confirmations_policy() {
         testing::pool::can_spend_inputs_by_confirmations_policy::<SaplingPoolTester>();
     }
+
+    #[test]
+    fn receive_two_notes_with_same_value() {
+        testing::pool::receive_two_notes_with_same_value::<SaplingPoolTester>();
+    }
 }